Add 'About' view to the SC Kill Monitor application
This commit introduces the 'About' view to the SC Kill Monitor application. This new view, managed by the AboutViewController, provides detailed app information, including its purpose, licensing details, and disclaimers. It also displays the app's logo. An 'About' button has been added to the MainViewController, which triggers the display of the 'About' view when pressed. The corresponding FXML file and logo have been added to the resources.

Closes #8

Signed-off-by: Lucas Greuloch (greluc) <lucas.greuloch@pm.me>